requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
dlib==19.24.99
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.1
psycopg2-binary==2.9.5
pytz==2022.7
yarl==1.8.2
